# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM Pay API",
        "description": "API de paiement mobile OM Pay",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000/api",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/auth/create-account": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Créer un nouveau compte avec token ou via authentification",
                "operationId": "010a15987b38f9aea8cd81e56f664d1d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "required": [
                                            "token",
                                            "name",
                                            "pin"
                                        ],
                                        "properties": {
                                            "token": {
                                                "description": "Token de création de compte",
                                                "type": "string",
                                                "example": "dy5l0kIMH1XV5SqKA8RLFpBxDMiTFcaimPJxmhYk5UlMEEoQWUsS97Oaiw6S"
                                            },
                                            "name": {
                                                "description": "Nom de l'utilisateur",
                                                "type": "string",
                                                "example": "Test User"
                                            },
                                            "pin": {
                                                "description": "PIN à 4 chiffres",
                                                "type": "string",
                                                "example": "1234"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "required": [
                                            "telephone",
                                            "pin"
                                        ],
                                        "properties": {
                                            "telephone": {
                                                "description": "Numéro de téléphone",
                                                "type": "string",
                                                "example": "+221776525959"
                                            },
                                            "pin": {
                                                "description": "PIN à 4 chiffres",
                                                "type": "string",
                                                "example": "1234"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte créé avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation ou token invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Token invalide ou expiré"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/generate-qr-code": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Générer un QR code pour les paiements",
                "operationId": "dfddbed7dcab63d0ee29f2afa715b572",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "compte_id"
                                ],
                                "properties": {
                                    "compte_id": {
                                        "description": "ID du compte",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "019a6d6f-7770-70a0-a293-cd767103d731"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "QR code généré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "QR code généré"
                                        },
                                        "data": {
                                            "properties": {
                                                "qr_code": {
                                                    "description": "Données du QR code encodées",
                                                    "type": "string"
                                                },
                                                "expires_at": {
                                                    "description": "Date d'expiration",
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès non autorisé au compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accès non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/get-solde": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Récupérer le solde d'un compte",
                "operationId": "5a555acc9bbfc489b1df78db209be801",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "compte_id"
                                ],
                                "properties": {
                                    "compte_id": {
                                        "description": "ID du compte",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "019a6d6f-7770-70a0-a293-cd767103d731"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Solde récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Solde récupéré avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "solde": {
                                                    "description": "Solde du compte",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1500.5
                                                },
                                                "compte_id": {
                                                    "description": "ID du compte",
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "019a6d6f-7770-70a0-a293-cd767103d731"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès non autorisé au compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accès non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Compte": {
                "title": "Compte",
                "description": "Représentation d'un compte utilisateur",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "db0f5a30-f7a8-4036-a2c1-5f6b0b1118c9"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+221776525959"
                    },
                    "solde": {
                        "type": "number",
                        "format": "float",
                        "example": 224900
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}